add native buy test 20240610

+++ Contracts/libraries/L_BalanceUINT256.sol
@@ -61,16 +61,18 @@ function lowerprice(
 ) pure returns (bool) {
     if (a.amount1() == 0 || b.amount0() == 0 || c.amount0() == 0) return false;
     require(
-        a.amount0() * b.amount1() * c.amount1() <= type(uint256).max,
+        uint256(a.amount0()) * uint256(b.amount1()) * uint256(c.amount1()) <=
+            type(uint256).max,
         "overflow"
     );
     require(
-        a.amount1() * b.amount0() * c.amount0() <= type(uint256).max,
+        uint256(a.amount1()) * uint256(b.amount0()) * uint256(c.amount0()) <=
+            type(uint256).max,
         "overflow"
     );
     return
-        a.amount0() * b.amount1() * c.amount1() <
-            a.amount1() * b.amount0() * c.amount0()
+        uint256(a.amount0()) * uint256(b.amount1()) * uint256(c.amount1()) <
+            uint256(a.amount1()) * uint256(b.amount0()) * uint256(c.amount0())
             ? true
             : false;
     //return uint256((a.amount0()*b.amount1())/(a.amount1()*b.amount0()));
